Upgrading SQL to SF-v4.1.0

+++ src/SoundFingerprinting.SQL/DAO/SubFingerprintDTO.cs
@@ -10,31 +10,31 @@
 
         public double SequenceAt { get; set; }
 
-        public long HashTable_0 { get; set; }
-        public long HashTable_1 { get; set; }
-        public long HashTable_2 { get; set; }
-        public long HashTable_3 { get; set; }
-        public long HashTable_4 { get; set; }
-        public long HashTable_5 { get; set; }
-        public long HashTable_6 { get; set; }
-        public long HashTable_7 { get; set; }
-        public long HashTable_8 { get; set; }
-        public long HashTable_9 { get; set; }
-        public long HashTable_10 { get; set; }
-        public long HashTable_11 { get; set; }
-        public long HashTable_12 { get; set; }
-        public long HashTable_13 { get; set; }
-        public long HashTable_14 { get; set; }
-        public long HashTable_15 { get; set; }
-        public long HashTable_16 { get; set; }
-        public long HashTable_17 { get; set; }
-        public long HashTable_18 { get; set; }
-        public long HashTable_19 { get; set; }
-        public long HashTable_20 { get; set; }
-        public long HashTable_21 { get; set; }
-        public long HashTable_22 { get; set; }
-        public long HashTable_23 { get; set; }
-        public long HashTable_24 { get; set; }
+        public int HashTable_0 { get; set; }
+        public int HashTable_1 { get; set; }
+        public int HashTable_2 { get; set; }
+        public int HashTable_3 { get; set; }
+        public int HashTable_4 { get; set; }
+        public int HashTable_5 { get; set; }
+        public int HashTable_6 { get; set; }
+        public int HashTable_7 { get; set; }
+        public int HashTable_8 { get; set; }
+        public int HashTable_9 { get; set; }
+        public int HashTable_10 { get; set; }
+        public int HashTable_11 { get; set; }
+        public int HashTable_12 { get; set; }
+        public int HashTable_13 { get; set; }
+        public int HashTable_14 { get; set; }
+        public int HashTable_15 { get; set; }
+        public int HashTable_16 { get; set; }
+        public int HashTable_17 { get; set; }
+        public int HashTable_18 { get; set; }
+        public int HashTable_19 { get; set; }
+        public int HashTable_20 { get; set; }
+        public int HashTable_21 { get; set; }
+        public int HashTable_22 { get; set; }
+        public int HashTable_23 { get; set; }
+        public int HashTable_24 { get; set; }
 
         public string Clusters { get; set; }
     }
